Fixed activation warnings and errors
Finally fixed the activation issue with Timber, class-seo-central.php has the sitemap function that sets the action after_setup_theme that was being triggered at the wrong time. Now activation works with timber already installed.

Some warnings were appearing on the dashboard and file editor for a few of the php conditonals and setup

Wrapped the robots file_get_content with checks to avoid warnings

Adjusted the functions for settings to avoid errors when keys are unavailable

Fixed all the image pathing for seo-central-wp-lite

Updated readme